Move column metadata tests to their own suite.
