+++ src/it/scala/com/databricks/spark/redshift/RedshiftIntegrationSuite.scala
@@ -220,97 +220,6 @@ class RedshiftIntegrationSuite extends IntegrationSuiteBase {
     rdd1.count()
   }
 
-  test("configuring maxlength on string columns") {
-    val tableName = s"configuring_maxlength_on_string_column_$randomSuffix"
-    try {
-      val metadata = new MetadataBuilder().putLong("maxlength", 512).build()
-      val schema = StructType(
-        StructField("x", StringType, metadata = metadata) :: Nil)
-      write(sqlContext.createDataFrame(sc.parallelize(Seq(Row("a" * 512))), schema))
-        .option("dbtable", tableName)
-        .mode(SaveMode.ErrorIfExists)
-        .save()
-      assert(DefaultJDBCWrapper.tableExists(conn, tableName))
-      checkAnswer(read.option("dbtable", tableName).load(), Seq(Row("a" * 512)))
-      // This append should fail due to the string being longer than the maxlength
-      intercept[SQLException] {
-        write(sqlContext.createDataFrame(sc.parallelize(Seq(Row("a" * 513))), schema))
-          .option("dbtable", tableName)
-          .mode(SaveMode.Append)
-          .save()
-      }
-    } finally {
-      conn.prepareStatement(s"drop table if exists $tableName").executeUpdate()
-      conn.commit()
-    }
-  }
-
-  test("configuring compression on columns") {
-    val tableName = s"configuring_compression_on_columns_$randomSuffix"
-    try {
-      val metadata = new MetadataBuilder().putString("encoding", "LZO").build()
-      val schema = StructType(
-        StructField("x", StringType, metadata = metadata) :: Nil)
-      write(sqlContext.createDataFrame(sc.parallelize(Seq(Row("a" * 128))), schema))
-        .option("dbtable", tableName)
-        .mode(SaveMode.ErrorIfExists)
-        .save()
-      assert(DefaultJDBCWrapper.tableExists(conn, tableName))
-      checkAnswer(read.option("dbtable", tableName).load(), Seq(Row("a" * 128)))
-      val encodingDF = sqlContext.read
-        .format("jdbc")
-        .option("url", jdbcUrl)
-        .option("dbtable",
-          s"""(SELECT "column", lower(encoding) FROM pg_table_def WHERE tablename='$tableName')""")
-        .load()
-      checkAnswer(encodingDF, Seq(Row("x", "lzo")))
-    } finally {
-      conn.prepareStatement(s"drop table if exists $tableName").executeUpdate()
-      conn.commit()
-    }
-  }
-
-  test("configuring comments on columns") {
-    val tableName = s"configuring_comments_on_columns_$randomSuffix"
-    try {
-      val metadata = new MetadataBuilder().putString("description", "Hello Column").build()
-      val schema = StructType(
-        StructField("x", StringType, metadata = metadata) :: Nil)
-      write(sqlContext.createDataFrame(sc.parallelize(Seq(Row("a" * 128))), schema))
-        .option("dbtable", tableName)
-        .option("description", "Hello Table")
-        .mode(SaveMode.ErrorIfExists)
-        .save()
-      assert(DefaultJDBCWrapper.tableExists(conn, tableName))
-      checkAnswer(read.option("dbtable", tableName).load(), Seq(Row("a" * 128)))
-      val tableDF = sqlContext.read
-        .format("jdbc")
-        .option("url", jdbcUrl)
-        .option("dbtable", s"(SELECT pg_catalog.obj_description('$tableName'::regclass))")
-        .load()
-      checkAnswer(tableDF, Seq(Row("Hello Table")))
-      val commentQuery =
-        s"""
-           |(SELECT c.column_name, pgd.description
-           |FROM pg_catalog.pg_statio_all_tables st
-           |INNER JOIN pg_catalog.pg_description pgd
-           |   ON (pgd.objoid=st.relid)
-           |INNER JOIN information_schema.columns c
-           |   ON (pgd.objsubid=c.ordinal_position AND c.table_name=st.relname)
-           |WHERE c.table_name='$tableName')
-         """.stripMargin
-      val columnDF = sqlContext.read
-        .format("jdbc")
-        .option("url", jdbcUrl)
-        .option("dbtable", commentQuery)
-        .load()
-      checkAnswer(columnDF, Seq(Row("x", "Hello Column")))
-    } finally {
-      conn.prepareStatement(s"drop table if exists $tableName").executeUpdate()
-      conn.commit()
-    }
-  }
-
   test("informative error message when saving a table with string that is longer than max length") {
     val tableName = s"error_message_when_string_too_long_$randomSuffix"
     try {
